move layouts from ui to components lib
